add data processing fix styles
