+++ static/dialogConstructorDev/src/components/MainLayout.tsx
@@ -1,9 +1,6 @@
 import React, { useState, useEffect } from "react";
-import * as _ from "lodash";
 import { Application } from "./Application";
-
 import { DiagramModel } from "@projectstorm/react-diagrams";
-import { Point } from "@projectstorm/geometry";
 import { CanvasWidget } from "@projectstorm/react-canvas-core";
 import { CustomNodeModel } from "./Node/CustomNodeModel";
 import {
@@ -15,21 +12,13 @@ import {
   NodeControlPanel,
   ZoomControlButton,
 } from "./style";
-import { AdvancedLinkModel } from "./LinksSettings";
+import { BackendGraphNode, dataForPost, receivedData } from './utils/dataProcessing';
 
 export interface MainLayoutProps {
   app: Application;
 }
 
-interface BackendGraphNode {
-  node_type: string;
-  node_name: string;
-  node_content: string;
-  node_links: string[];
-  node_views: number;
-  position_x: number;
-  position_y: number;
-}
+
 
 export interface BackendGraph {
   id: string;
@@ -43,8 +32,8 @@ export interface BackendShortGraph {
 }
 
 const baseUrl = location.host.startsWith("localhost:34567")
-  ? "http://localhost:62544/"
-  : location.origin + "/";
+    ? "http://localhost:62544/"
+    : location.origin + "/";
 
 const getListOfGraphs = async (): Promise<BackendShortGraph[]> => {
   const req = await fetch(baseUrl + "api/v1/dialo_graph.list");
@@ -75,7 +64,7 @@ const addNewGraph = async (graphData: BackendGraph): Promise<BackendGraph> => {
 };
 
 const updateGraphInBD = async (
-  graphData: BackendGraph
+    graphData: BackendGraph
 ): Promise<BackendGraph> => {
   const response = await fetch(baseUrl + "api/v1/dialo_graph.update", {
     method: "POST",
@@ -109,39 +98,9 @@ export const MainLayout = ({ app }: MainLayoutProps) => {
   };
 
   const saveGraph = async () => {
-    const serializedData = app.diagramEngine.getModel().serialize();
-    const nodeData = Object.values(serializedData.layers[1].models);
-    const arrowsData = Object.values(serializedData.layers[0].models);
-
-    const names = nodeData.map((node) => {
-      return { id: node.id, name: node.extras.name };
-    });
-
-    const momChild = arrowsData.map((arrow) => {
-      return {
-        mom: arrow.source,
-        child: arrow.target,
-        childName: names.find((name) => name.id === arrow.target)?.name,
-      };
-    });
-
-    const graph = nodeData.map((node) => {
-      return {
-        node_type: node.extras.type,
-        node_name: node.extras.name,
-        node_content: node.extras.content,
-        node_links: momChild
-          .filter((item) => item.mom === node.id)
-          .map((item) => item.childName),
-        node_views: 0,
-        position_x: node.x,
-        position_y: node.y,
-      };
-    });
-
+    const graph = dataForPost(app.diagramEngine.getModel());
     const graphData = await getOneGraph(selectGraph.id);
     const isNewGraph = !!graphData.error;
-
     const id = selectGraph.id;
 
     let title = selectGraph.title;
@@ -171,7 +130,7 @@ export const MainLayout = ({ app }: MainLayoutProps) => {
       console.log(e);
     }
 
-    updateListOfDialogs();
+    await updateListOfDialogs();
   };
 
   const updateListOfDialogs = async () => {
@@ -198,87 +157,18 @@ export const MainLayout = ({ app }: MainLayoutProps) => {
     if (window.confirm("Are you sure you want to change Dialog?")) {
       const graphFullData = await getOneGraph(newGraphId);
       const backendNodes = graphFullData.graph || [];
-
-      const intents = backendNodes.filter(
-        (backendNode) => backendNode.node_type === "intent"
-      );
-      const skills = backendNodes.filter(
-        (backendNode) => backendNode.node_type === "skill"
-      );
-
-      const allModels = [];
-      const intentModels = intents.map((backendNode) => {
-        const nodeModel = new CustomNodeModel(
-          backendNode.node_name,
-          backendNode.node_type,
-          false,
-          backendNode.node_content
-        );
-
-        nodeModel.setPosition(
-          new Point(backendNode.position_x, backendNode.position_y)
-        );
-        allModels.push(nodeModel);
-        return nodeModel;
-      });
-
-      const skillsModels = skills.map((backendNode) => {
-        const nodeModel = new CustomNodeModel(
-          backendNode.node_name,
-          backendNode.node_type,
-          true,
-          backendNode.node_content
-        );
-
-        nodeModel.setPosition(
-          new Point(backendNode.position_x, backendNode.position_y)
-        );
-        allModels.push(nodeModel);
-        return nodeModel;
-      });
-
-      skills.map((skillBackendNode, index) => {
-        skillBackendNode.node_links.map((nodeLink) => {
-          const skillModel = skillsModels[index];
-          intentModels
-            .filter((node) => node.getName() === nodeLink)
-            .forEach((intentModel) => {
-              const link = new AdvancedLinkModel();
-              link.setSourcePort(skillModel.getPort("in"));
-              link.setTargetPort(intentModel.getPort("out"));
-              allModels.push(link);
-            });
-        });
-      });
-
-      intents.forEach((intentBackendNode, index) => {
-        intentBackendNode.node_links.map((nodeLink) => {
-          const intentModel = intentModels[index];
-
-          skillsModels
-            .filter((node) => node.getName() === nodeLink)
-            .map((skillModel) => {
-              const link = new AdvancedLinkModel();
-              link.setSourcePort(intentModel.getPort("out"));
-              link.setTargetPort(skillModel.getPort("in"));
-              allModels.push(link);
-            });
-        });
-      });
-
+      const newModels = receivedData(backendNodes);
       app.activeModel = new DiagramModel();
       app.diagramEngine.setModel(app.activeModel);
-      app.activeModel.addAll(...allModels);
-
+      app.activeModel.addAll(...newModels);
       app.diagramEngine.repaintCanvas();
-
       setSelectGraph(graphFullData);
     }
   };
 
   const newGraph = async () => {
     const confirmMessage =
-      "Создать новый диалог? Введенные данные будут утеряны";
+        "Создать новый диалог? Введенные данные будут утеряны";
 
     if (selectGraph && !window.confirm(confirmMessage)) {
       return;
@@ -301,103 +191,103 @@ export const MainLayout = ({ app }: MainLayoutProps) => {
   };
 
   return (
-    <div>
-      <div
-        style={{
-          height: "100vh",
-        }}
-        onDrop={dropHandler}
-        onDragOver={(event) => {
-          event.preventDefault();
-        }}
-      >
+      <div>
+        <div
+            style={{
+              height: "100vh",
+            }}
+            onDrop={dropHandler}
+            onDragOver={(event) => {
+              event.preventDefault();
+            }}
+        >
+          {selectGraph && (
+              <GraphCanvas
+                  color="rgb(222, 222, 222)"
+                  background="rgb(233, 233, 233)"
+              >
+                <CanvasWidget engine={app.diagramEngine} />
+              </GraphCanvas>
+          )}
+        </div>
+
+        <DialogConstructorHeader>
+          {graphList.length > 0 && (
+              <select
+                  name="dialogs"
+                  value={selectGraph?.id}
+                  onChange={(e) => {
+                    changeGraph(e.target.value);
+                  }}
+              >
+                <option value="0"></option>
+                {graphList.map((graphInfo, key) => {
+                  return (
+                      <option value={graphInfo.id} key={graphInfo.id}>
+                        {graphInfo.title}
+                      </option>
+                  );
+                })}
+              </select>
+          )}
+
+          <AddDialogButton onClick={newGraph}>
+            Добавить новый диалог
+          </AddDialogButton>
+
+          {selectGraph && (
+              <AddDialogButton onClick={saveGraph}>Сохранить граф</AddDialogButton>
+          )}
+        </DialogConstructorHeader>
+
         {selectGraph && (
-          <GraphCanvas
-            color="rgb(222, 222, 222)"
-            background="rgb(233, 233, 233)"
-          >
-            <CanvasWidget engine={app.diagramEngine} />
-          </GraphCanvas>
+            <>
+              <NodeControlPanel>
+                <NodeControlElement
+                    color="#1A1A4E"
+                    draggable={true}
+                    onDragStart={(event) => {
+                      event.dataTransfer.setData("storm-diagram-node-type", "intent");
+                    }}
+                    className="tray-item"
+                >
+                  Intent
+                </NodeControlElement>
+
+                <NodeControlElement
+                    color="#FF7A00"
+                    draggable={true}
+                    onDragStart={(event) => {
+                      event.dataTransfer.setData("storm-diagram-node-type", "skill");
+                    }}
+                    className="tray-item"
+                >
+                  Skill
+                </NodeControlElement>
+              </NodeControlPanel>
+
+              <ZoomControlButton
+                  onClick={() => {
+                    app.diagramEngine.zoomToFitSelectedNodes({ margin: 100 });
+                  }}
+              >
+                Zoom to fit
+              </ZoomControlButton>
+
+              <DeleteControlButton
+                  onClick={() => {
+                    const confirm = window.confirm(
+                        "Are you sure you want to delete Dialog?"
+                    );
+                    if (confirm) {
+                      console.log("Send request");
+                    }
+                  }}
+              >
+                🗑
+              </DeleteControlButton>
+            </>
         )}
       </div>
-
-      <DialogConstructorHeader>
-        {graphList.length > 0 && (
-          <select
-            name="dialogs"
-            value={selectGraph?.id}
-            onChange={(e) => {
-              changeGraph(e.target.value);
-            }}
-          >
-            <option value="0"></option>
-            {graphList.map((graphInfo, key) => {
-              return (
-                <option value={graphInfo.id} key={graphInfo.id}>
-                  {graphInfo.title}
-                </option>
-              );
-            })}
-          </select>
-        )}
-
-        <AddDialogButton onClick={newGraph}>
-          Добавить новый диалог
-        </AddDialogButton>
-
-        {selectGraph && (
-          <AddDialogButton onClick={saveGraph}>Сохранить граф</AddDialogButton>
-        )}
-      </DialogConstructorHeader>
-
-      {selectGraph && (
-        <>
-          <NodeControlPanel>
-            <NodeControlElement
-              color="#1A1A4E"
-              draggable={true}
-              onDragStart={(event) => {
-                event.dataTransfer.setData("storm-diagram-node-type", "intent");
-              }}
-              className="tray-item"
-            >
-              Intent
-            </NodeControlElement>
-
-            <NodeControlElement
-              color="#FF7A00"
-              draggable={true}
-              onDragStart={(event) => {
-                event.dataTransfer.setData("storm-diagram-node-type", "skill");
-              }}
-              className="tray-item"
-            >
-              Skill
-            </NodeControlElement>
-          </NodeControlPanel>
-
-          <ZoomControlButton
-            onClick={() => {
-              app.diagramEngine.zoomToFitSelectedNodes({ margin: 100 });
-            }}
-          >
-            Zoom to fit
-          </ZoomControlButton>
-
-          <DeleteControlButton
-            onClick={() => {
-              const confirm = window.confirm(
-                "Are you sure you want to delete Dialog?"
-              );
-              if (confirm) {
-                console.log("Send request");
-              }
-            }}
-          >
-            🗑
-          </DeleteControlButton>
-        </>
-      )}
-    </div>
   );
 };
